Add auto-heal role
This patch adds a new `openshift_autoheal` role that creates an example
of the auto-heal solution.  Currently the only component of this
solution is the `autoheal` deployment, which runs a server that accepts
Prometheus alert manager notifications and triggers the exeuction of AWX
(a.k.a. Ansible Tower) jobs.

To use the role create and run a playbook like this:

```yaml
---
- hosts: master
  remote_user: root
  vars:
    openshift_autoheal_alert_pattern: "MyAlert"
    openshift_autoheal_awx_host: "myawx.example.com"
    openshift_autoheal_awx_port: "443"
    openshift_autoheal_awx_user: "myawxuser"
    openshift_autoheal_awx_password: "..."
    openshift_autoheal_awx_project: "My AWX project"
    openshift_autoheal_awx_template: "My AWX template"
  roles:
  - openshift_autoheal
```

Note that the `openshift_autoheal_alert_pattern` variable should contain
a regular expression. Only the alerts whose `alertname` label match this
regular expression will be processed by the auto-heal components.

By default all the components are deployed in a new `openshift-autoheal`
namespace.

To check if the deployment is working correctly take a look at the looks
of the `autoheal-...` pods:

```
$ oc get pod -n openshift-autoheal
```

Then check the logs of the pod:

```
$ oc logs -n openshift-autoheal autoheal-... -f
```

Initially it should print a line like this:

```
Listening on address '0.0.0.0' and port 9099
```

To use it add a configuration like this to the Prometheus alert manager
configuration, so that it will send alert notifications to this
component:

```yaml
receivers:
- name: autoheal
  webhook_configs:
  - url: http://autoheal.openshift-autoheal.svc.cluster.local:9099
```

Then configure the Prometheus alert manager rules so that the alerts
that you need are sent to this `autoheal` receiver.

When an alert is received you will see this in the log:

```
Alert manager request: ...
Alert 'MyAlert' matches pattern 'MyAlert'
...
```

And after that you will see all the details of the communication with
AWX.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>