[WIP] Proposal for auto-heal rule definition contract
This patch shows a proposal of how the contract for defining auto-heal
rules could work. The idea is creating a new `authoheal_rule` Ansible
module that hides the details of how the auto-heal rules are technically
defined. The rules could be defined as instances of a Kubernetes custom
resource definition, or the could be defined as fragmets of a YAML file,
or as requests to remote REST server, or as messages sent to a queue,
but the developer of the rule does not need to care about that.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>