Add remaining Tectonic scrapers and alerts
Bring in the remaining built-in Tectonic service scrapers and alerting rules.
Add a script which allows the Origin controller manager to be scraped (TODO:
doesn't work with `oc cluster up` clusters yet).

Note that the scheduler is currently part of the Origin controller manager and
so scheduler metrics are captured under the kube-controller-manager job for now.
This means the default scheduler alerts will be broken. This isn't worth fixing
right now given the plans for Origin in 3.10 and the ongoing Tectonic Prometheus
Operator work.